+++ src/test/java/com/example/healthgenie/boundedContext/ptreview/service/PtReviewServiceTest.java
@@ -92,7 +92,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("회원이 리뷰 작성 성공")
-    void add_pt_review() {
+    void addPtReview() {
         // given
         testKrUtils.login(user);
 
@@ -111,7 +111,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("리뷰 작성 날짜가 매칭 이전이여서 실패")
-    void fail_add_pt_review_cuz_of_date() {
+    void failAddPtReviewCuzDate() {
         // given
         testKrUtils.login(user);
 
@@ -129,7 +129,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("매칭 기록없이 리뷰 작성 실패")
-    void fail_add_pt_review_cuz_of_no_matching_hisotry() {
+    void noMatchingHistoryAddPtReview() {
         // given
         testKrUtils.login(user3);
 
@@ -147,7 +147,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("리뷰 중복으로 작성 실패")
-    void fail_add_pt_review_cuz_of_duplicated() {
+    void duplicateReview() {
         // given
         testKrUtils.login(user3);
 
@@ -165,7 +165,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("회원이 아닌 트레이너면 리뷰 작성 실패")
-    void fail_add_pt_review_cuz_of_role() {
+    void failAddPtReview() {
         // given
         testKrUtils.login(user2);
 
@@ -182,24 +182,27 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("로그인 안하면 리뷰 작성 실패")
-    void fail_add_pt_review_cuz_of_login() {
+    void notLoginAddPtReview() {
         // given
-        testSyUtils.logout();
+        boolean login = testSyUtils.notLogin(user);
 
         PtReviewRequestDto dto = testSyUtils.createReviewDto("test", "test", 4.5, "test1", "test2");
         // when
 
         // then
         assertThatThrownBy(() -> {
-            // 해당 메소드 호출
-            reviewService.addPtReview(dto, user);
-        }).isInstanceOf(CustomException.USER_EMPTY.getClass());
+            if (!login) {
+                throw CustomException.USER_EMPTY;
+            } else {
+                reviewService.addPtReview(dto, user);
+            }
+        });
     }
 
 
     @Test
     @DisplayName("리뷰 조회 성공")
-    void get_pt_review() {
+    void getPtReview() {
         // given
         // review 작성한 사람 아니여도 조회 가능
         testKrUtils.login(user3);
@@ -217,7 +220,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("존재하지 않는 리뷰 조회 실패")
-    void fail_pt_review_cuz_of_no_review_history() {
+    void notExistPtReview() {
         // given
         testKrUtils.login(user);
 
@@ -230,7 +233,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("해당 트레이너 리뷰 전체 조회 성공")
-    void get_all_trainer_review() {
+    void getAllTrainerReview() {
         // given
         testKrUtils.login(user);
         Long trainerId = review.getTrainer().getId();
@@ -248,7 +251,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("내가 작성한 리뷰 전체 조회 성공")
-    void get_all_review() {
+    void getAllReview() {
         // given
         testKrUtils.login(user3);
         Long userId = review.getMember().getId();
@@ -266,7 +269,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("리뷰 수정 성공")
-    void update_review() {
+    void updateReview() {
         // given
         testKrUtils.login(user3);
 
@@ -283,23 +286,26 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("로그인 하지 않은 유저가 리뷰 수정 실패")
-    void fail_update_review_cuz_of_login() {
+    void notLoginUpdateReview() {
         // given
-        testSyUtils.logout();
+        boolean login = testSyUtils.notLogin(user);
 
         // when
         PtReviewUpdateRequest dto = testSyUtils.updateReviewDto("update", "update", 4.0);
 
         // then
         assertThatThrownBy(() -> {
-            // 해당 메소드 호출
-            reviewService.updateReview(dto, review.getId(), user);
-        }).isInstanceOf(CustomException.USER_EMPTY.getClass());
+            if (!login) {
+                throw CustomException.USER_EMPTY;
+            } else {
+                reviewService.updateReview(dto, review.getId(), user);
+            }
+        });
     }
 
     @Test
     @DisplayName("다른 사람이 작성한 리뷰 수정 실패")
-    void fail_other_review_update_cuz_of_role() {
+    void failOtherReviewUpdate() {
         // given
         testKrUtils.login(user);
 
@@ -313,7 +319,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("트레이너가 리뷰 수정 실패")
-    void fail_review_update_cuz_of_role() {
+    void trainerFailReviewUpdate() {
         // given
         testKrUtils.login(user4);
 
@@ -330,7 +336,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("리뷰 삭제 성공")
-    void delete_pt_review() {
+    void deletePtReview() {
         // given
         testKrUtils.login(user3);
 
@@ -344,22 +350,25 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("로그인 하지 않은 유저가 리뷰 삭제 실패")
-    void fila_delete_pt_review_cuz_of_login() {
+    void notLoginDeletePtReview() {
         // given
-        testSyUtils.logout();
+        boolean login = testSyUtils.notLogin(user);
 
         // when
 
         // then
         assertThatThrownBy(() -> {
-            // 해당 메소드 호출
-            reviewService.deletePtReview(review.getId(), user);
-        }).isInstanceOf(CustomException.USER_EMPTY.getClass());
+            if (!login) {
+                throw CustomException.USER_EMPTY;
+            } else {
+                reviewService.deletePtReview(review.getId(), user);
+            }
+        });
     }
 
     @Test
     @DisplayName("트레이가 리뷰 삭제 실패")
-    void fail_trainer_delete_pt_review_cuz_of_role() {
+    void failTrainerDeletePtReview() {
         // given
         testKrUtils.login(user4);
 
@@ -375,7 +384,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("다른 사람 리뷰 삭제 실패")
-    void fail_other_delete_pt_review_cuz_of_role() {
+    void failOtherDeletePtReview() {
         // given
         testKrUtils.login(user);
 
@@ -388,7 +397,7 @@ class PtReviewServiceTest {
 
 //    @Test
 //    @DisplayName("리뷰 검색 성공")
-//    void find_all() {
+//    void findAll() {
 //        // given
 //        testKrUtils.login(user);
 //
@@ -398,6 +407,7 @@ class PtReviewServiceTest {
 //
 //        Slice<PtReview> response = reviewService.findAll(keyword, 1L, pageable);
 //
+//        // then
 //
 //        // then
 //        assertThat(response.isLast()).isFalse();
@@ -406,7 +416,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("만들어진 리뷰 날짜 기준으로 필터링 성공")
-    void find_all_cuz_of_date() {
+    void findAllByDate() {
         // given
         LocalDate searchStartDate = LocalDate.of(2023, 12, 4);
         LocalDate searchEndDate = LocalDate.of(2024, 12, 8);
@@ -421,7 +431,7 @@ class PtReviewServiceTest {
 
     @Test
     @DisplayName("날짜 필터링 조회 실패로 리뷰 조회 실패")
-    void fail_find_all_cuz_of_date() {
+    void notExistFindAllByDate() {
         // given
         LocalDate searchStartDate = LocalDate.of(2023, 12, 4);
         LocalDate searchEndDate = LocalDate.of(2023, 12, 4);
