+++ src/test/java/com/example/healthgenie/boundedContext/routine/service/RoutineServiceTest.java
@@ -41,6 +41,7 @@ class RoutineServiceTest {
     RoutineService routineService;
 
     User user;
+    User user1;
     Routine routine;
     Routine beginnerGenie;
     Routine intermediateGenie;
@@ -58,6 +59,7 @@ class RoutineServiceTest {
 
         user = testKrUtils.createUser("jh485200@gmail.com", "test1", AuthProvider.EMPTY, Role.USER);
         routine = testSyUtils.writeRoutine(Day.WEDNESDAY, "하체,가슴", workoutRecipeList, user);
+        //routine = testSyUtils.writeRoutine(Day.WEDNESDAY, "하체,가슴", recipe, user);
 
         beginnerGenie = testSyUtils.genieRoutine(Level.BEGINNER, Day.FRIDAY, "test", "하체,가슴", recipe);
         intermediateGenie = testSyUtils.genieRoutine(Level.INTERMEDIATE, Day.FRIDAY, "test", "하체,가슴", recipe);
@@ -72,7 +74,7 @@ class RoutineServiceTest {
 
     @Test
     @DisplayName("루틴 작성하기")
-    void write_routine() {
+    void writeRoutine() {
         // given
         testKrUtils.login(user);
 
@@ -105,7 +107,7 @@ class RoutineServiceTest {
 
     @Test
     @DisplayName("중복된 요일 루틴 작성 실패")
-    void fail_add_routine_cuz_of_duplicated_day() {
+    void failAddRoutineCuzDuplicateDay() {
         // given
         testKrUtils.login(user);
 
@@ -127,9 +129,9 @@ class RoutineServiceTest {
 
     @Test
     @DisplayName("로그인 하지 않은 유저가 루틴 작성하기")
-    void fail_write_routine_cuz_of_login() {
+    void notLoginWriteRoutine() {
         // given
-        testSyUtils.logout();
+        boolean loginResult = testSyUtils.notLogin(user);
 
         WorkoutRecipe recipe = new WorkoutRecipe("스쿼트", 3, 3, 3);
 
@@ -140,15 +142,18 @@ class RoutineServiceTest {
 
         // then
         assertThatThrownBy(() -> {
-            // 해당 메소드 호출
-            routineService.writeRoutine(dto, user);
-        }).isInstanceOf(CustomException.UNKNOWN_EXCEPTION.getClass());
+            if (!loginResult) {
+                throw CustomException.UNKNOWN_EXCEPTION;
+            } else {
+                routineService.writeRoutine(dto, user);
+            }
+        }).isInstanceOf(CustomException.class);
     }
 
 
     @Test
     @DisplayName("루틴 수정하기")
-    void update_routine() {
+    void updateRoutine() {
 
         /*
             본인 루틴은 본인만 보기에 타인이 수정하거나 삭제할 일이 없다.
@@ -187,7 +192,7 @@ class RoutineServiceTest {
 
     @Test
     @DisplayName("나의 루틴 모두 조회하기")
-    void get_all_my_routine() {
+    void getAllMyRoutine() {
         // given
         testKrUtils.login(user);
 
@@ -212,7 +217,7 @@ class RoutineServiceTest {
 
     @Test
     @DisplayName("나의 루틴 요일별 상세조회")
-    void get_my_routine() {
+    void getMyRoutine() {
         // given
         testKrUtils.login(user);
 
@@ -230,7 +235,7 @@ class RoutineServiceTest {
 
     @Test
     @DisplayName("level 별 지니 루틴 모두 조회하기")
-    void get_all_genie_routine() {
+    void getAllGenieRoutine() {
         // given
         testKrUtils.login(user);
 
@@ -254,7 +259,7 @@ class RoutineServiceTest {
 
     @Test
     @DisplayName("level이 empty면 지니 루틴 조회 실패")
-    void fail_genie_routine_cuz_of_empty_level() {
+    void failGenieRoutine() {
         // given
         testKrUtils.login(user);
 
@@ -271,7 +276,7 @@ class RoutineServiceTest {
 
     @Test
     @DisplayName("해당 요일의 지니 루틴 조회하기")
-    void get_genie_routine() {
+    void getGenieRoutine() {
         // given
         List<RoutineResponseDto> begin = routineService.getGenieRoutine(Level.BEGINNER, Day.FRIDAY);
         List<RoutineResponseDto> inter = routineService.getGenieRoutine(Level.BEGINNER, Day.FRIDAY);
@@ -292,7 +297,7 @@ class RoutineServiceTest {
 
     @Test
     @DisplayName("루틴 삭제하기")
-    void delete_routine() {
+    void deleteRoutine() {
         // given
         testKrUtils.login(user);
 
@@ -306,22 +311,25 @@ class RoutineServiceTest {
 
     @Test
     @DisplayName("로그인 하지 않은 유저가 루틴 삭제하기")
-    void fail_delete_routine_cuz_of_login() {
+    void notLogindeleteRoutine() {
         // given
-        testSyUtils.logout();
+        boolean loginResult = testSyUtils.notLogin(user);
 
         // when
 
         // then
         assertThatThrownBy(() -> {
-            // 해당 메소드 호출
-            routineService.deleteRoutine(routine.getId(), user);
-        }).isInstanceOf(CustomException.USER_EMPTY.getClass());
+            if (!loginResult) {
+                throw CustomException.UNKNOWN_EXCEPTION;
+            } else {
+                routineService.deleteRoutine(routine.getId(), user);
+            }
+        }).isInstanceOf(CustomException.class);
     }
 
     @Test
     @DisplayName("존재 하지 않은 루틴 삭제하기")
-    void fail_routine_delete_cuz_of_no_routine_history() {
+    void notExistRoutineDelete() {
         // given
         testKrUtils.login(user);
