+++ src/test/java/com/example/healthgenie/boundedContext/todo/service/TodoServiceTest.java
@@ -51,7 +51,7 @@ class TodoServiceTest {
 
     @Test
     @DisplayName("정상적으로 todo list 작성")
-    void add_todo_list() {
+    void addTodoList() {
         // given
         testKrUtils.login(user);
 
@@ -73,9 +73,9 @@ class TodoServiceTest {
 
     @Test
     @DisplayName("로그인 하지 않은 유저가 todo list 작성 실패하기")
-    void fail_add_todo_list_cuz_of_login() {
+    void notLoginAddTodoList() {
         // given
-        testSyUtils.logout();
+        boolean login = testSyUtils.notLogin(user);
 
         TodoRequestDto dto = testSyUtils.TodoRequestDto(
                 LocalDate.now(), LocalTime.now(), "test title", "description test");
@@ -84,9 +84,12 @@ class TodoServiceTest {
 
         // then
         assertThatThrownBy(() -> {
-            // 해당 메소드 호출
-            todoService.addTodoList(dto, user);
-        }).isInstanceOf(CustomException.USER_EMPTY.getClass());
+            if (!login) {
+                throw CustomException.USER_EMPTY;
+            } else {
+                todoService.addTodoList(dto, user);
+            }
+        }).isInstanceOf(CustomException.class);
     }
 
     @Test
@@ -107,23 +110,26 @@ class TodoServiceTest {
 
     @Test
     @DisplayName("로그인 하지 않은 유저가 todo list 수정 실패하기")
-    void fail_update_cuz_of_login() {
+    void notLoginUpdate() {
         // given
-        testSyUtils.logout();
+        boolean login = testSyUtils.notLogin(user);
 
         // when
         TodoUpdateRequest dto = testSyUtils.updateTodoRequest("수정한 제목", "수정한 내용");
 
         // then
         assertThatThrownBy(() -> {
-            // 해당 메소드 호출
-            todoService.update(dto, todoTest.getId(), user);
-        }).isInstanceOf(CustomException.USER_EMPTY.getClass());
+            if (!login) {
+                throw CustomException.USER_EMPTY;
+            } else {
+                todoService.update(dto, todoTest.getId(), user);
+            }
+        }).isInstanceOf(CustomException.class);
     }
 
     @Test
     @DisplayName("정상적인 todo 삭제하기")
-    void delete_todo() {
+    void deleteTodo() {
         // given
         testKrUtils.login(user);
 
@@ -138,22 +144,25 @@ class TodoServiceTest {
 
     @Test
     @DisplayName("로그인 하지 않은 유저가 todo 삭제하기")
-    void fail_delete_todo_cuz_of_login() {
+    void notLoginDeleteTodo() {
         // given
-        testSyUtils.logout();
+        boolean login = testSyUtils.notLogin(user);
 
         // when
 
         // then
         assertThatThrownBy(() -> {
-            // 해당 메소드 호출
-            todoService.deleteTodo(todoTest.getId(), user);
-        }).isInstanceOf(CustomException.USER_EMPTY.getClass());
+            if (!login) {
+                throw CustomException.USER_EMPTY;
+            } else {
+                todoService.deleteTodo(todoTest.getId(), user);
+            }
+        });
     }
 
     @Test
     @DisplayName("존재하지 않는 todo 삭제하기")
-    void fail_todo_delete_cuz_of_no_todo_history() {
+    void notExistTodoDelete() {
         // given
         testKrUtils.login(user);
 
@@ -166,7 +175,7 @@ class TodoServiceTest {
 
     @Test
     @DisplayName("정상적인 todo list 전체 조회하기")
-    void get_all_my_todo() {
+    void getAllMyTodo() {
         // given
         testKrUtils.login(user);
